pyproject.toml
[project]
name = "voice-memo-summarizer"
version = "0.1.0"
description = "Summarize Apple Voice Memos using Gemini and upload to Craft"
readme = "README.md"
authors = [
    { name = "Jeroen Overschie" }
]
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.67.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[project.scripts]
voice-memo-summarizer = "voice_memo_summarizer.__main__:main"

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
